[#96] 무한호출 오류 해결 draft (공유용)
