+++ src/components/FeedCard.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 import FeedCardBadge from './FeedCardBadge'
 import FeedCardQuestion from './FeedCardQuestion'
 import FeedCardAnswer from './FeedCardAnswer'
@@ -6,18 +6,25 @@ import FeedCardAnswerEdit from './FeedCardAnswerEdit'
 import FeedCardReaction from './FeedCardReaction'
 import styles from './FeedCard.module.css'
 import { MenuDropdown, MenuDropdownItem } from './Dropdown'
+import deleteFeedQuestion from '../api/deleteFeedQuestion'
+import Toast from './Toast'
 
 export default function FeedCard({
   question,
   isMyFeed,
   profile = null,
-  onDeleteClick,
   onLoad,
 }) {
   const [isEditing, setIsEditing] = useState(false)
+  const [errorInfo, setErrorInfo] = useState(null)
 
-  const handleDeleteClick = () => {
-    onDeleteClick(question.id)
+  const handleDeleteClick = async () => {
+    try {
+      await deleteFeedQuestion(question.id)
+    } catch (error) {
+      setErrorInfo(error.message)
+    }
+    await onLoad()
   }
 
   const handleEditClick = () => {
@@ -28,8 +35,12 @@ export default function FeedCard({
 
   const submitEdit = () => {
     setIsEditing(false)
-    onLoad(question.subjectId)
+    onLoad()
   }
+
+  useEffect(() => {
+    onLoad()
+  }, [isEditing, onLoad])
 
   return (
     <div className={styles['card-container']}>
@@ -62,6 +73,7 @@ export default function FeedCard({
       )}
       <div className={styles['card-division-line']} />
       <FeedCardReaction question={question} />
+      {errorInfo && <Toast>{errorInfo}</Toast>}
     </div>
   )
 }
