+++ src/pages/IndividualFeedPage/IndividualFeedPage.jsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { Link, useParams } from 'react-router-dom'
 import FeedCardList from '../../components/FeedCardList'
 import SocialShareContainer from '../../components/SocialShareContainer'
@@ -6,7 +6,6 @@ import { FloatButton } from '../../components/Buttons'
 import QuestionModal from '../../components/QuestionModal/QuestionModal'
 import getProfileById from '../../api/getProfileById'
 import Toast from '../../components/Toast'
-import getFeedQuestions from '../../api/getFeedQuestions'
 import { getLocalUserId } from '../../modules/utils'
 import styles from './IndividualFeedPage.module.css'
 import openMindImg from '../../assets/images/img_openmind.png'
@@ -15,17 +14,9 @@ import logo from '../../assets/images/img_logo.png'
 export default function IndividualFeedPage() {
   const { feedId } = useParams()
   const [profile, setProfile] = useState(null)
-  const [questions, setQuestions] = useState([])
-  const [questionCount, setQuestionCount] = useState(0)
-  const [offset, setOffset] = useState(0)
-  const [next, setNext] = useState(null)
   const [errorInfo, setErrorInfo] = useState(null)
-  const [isLoading, setIsLoading] = useState(false)
-  const LIMIT = 10
 
   const localUserId = getLocalUserId()
-
-  const listEnd = document.querySelector('.list-end')
 
   const handleLoadProfile = async (id) => {
     let response
@@ -39,59 +30,9 @@ export default function IndividualFeedPage() {
     return null
   }
 
-  const handleLoadQuestions = useCallback(async () => {
-    if (isLoading) {
-      return
-    }
-    let response
-    try {
-      setIsLoading(true)
-      response = await getFeedQuestions({ feedId, offset, limit: LIMIT })
-      if (offset === 0) {
-        setQuestions(response.results)
-      } else {
-        setQuestions((prevQuestions) => [...prevQuestions, ...response.results])
-      }
-      setQuestionCount(response.count)
-      setOffset((prevOffset) => prevOffset + response.results.length)
-      setNext(response.next)
-      setErrorInfo(null)
-    } catch (error) {
-      setErrorInfo(error.message)
-    } finally {
-      setIsLoading(false)
-    }
-  }, [feedId, offset, isLoading])
-
   useEffect(() => {
     handleLoadProfile(feedId).then()
   }, [feedId])
-
-  useEffect(() => {
-    if (offset === 0) {
-      handleLoadQuestions().then()
-    }
-  }, [feedId, offset, handleLoadQuestions])
-
-  const observer = new IntersectionObserver(
-    (entries) => {
-      if (questionCount > 0 && entries[0].isIntersecting) {
-        observer.unobserve(listEnd)
-        handleLoadQuestions().then()
-      } else if (questionCount === 0) {
-        observer.disconnect()
-      }
-    },
-    { threshold: 0.5 }
-  )
-
-  if (listEnd) {
-    observer.observe(listEnd)
-  }
-
-  if (errorInfo) {
-    return <Toast>{errorInfo}</Toast>
-  }
 
   if (profile) {
     return (
@@ -113,23 +54,8 @@ export default function IndividualFeedPage() {
         </header>
         <main className={styles.main}>
           <section className={styles['feed-card-list']}>
-            <FeedCardList
-              questions={questions}
-              questionCount={questionCount}
-              isMyFeed={false}
-              profile={profile}
-              setErrorInfo={setErrorInfo}
-              onLoad={handleLoadQuestions}
-            />
+            <FeedCardList isMyFeed={false} profile={profile} />
           </section>
-          <div
-            className="list-end"
-            style={
-              questionCount === 0 || next === null ? { display: 'none' } : {}
-            }
-          >
-            로딩 중 ...
-          </div>
         </main>
         <div className={styles['button-floating']}>
           {localUserId === feedId ? ( // 로그인한 사용자와 프로필 주인이 같은 경우 답변하기 버튼을 보여줍니다.
@@ -137,7 +63,7 @@ export default function IndividualFeedPage() {
               <FloatButton>답변하기</FloatButton>
             </Link>
           ) : (
-            <QuestionModal profile={profile} onLoad={handleLoadQuestions} />
+            <QuestionModal profile={profile} />
           )}
         </div>
         {errorInfo && <Toast>{errorInfo}</Toast>}
